v0.2.20-feat: add TimeCNNEncoder and eTimeCNNEncoder modules
- Updated version in pyproject.toml to 0.2.20
- Added TimeCNNEncoder class for 1D CNN encoding
- Added eTimeCNNEncoder class for equivariant 1D CNN encoding
- Updated import paths in __init__.py for new modules
