+++ .github/workflows/tests.yaml
@@ -39,4 +39,4 @@ jobs:
 
       - name: Run Tests
         working-directory: ${{ github.workspace }}
-        run: pytest -s test/
+        run: pytest -s -v test/
