+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "symm-learning"
-version = "0.2.19"
+version = "0.2.20"
 
 description = "Torch modules and utilities of equivariant/invariant learning"
 readme = "README.md"
