+++ symm_learning/models/__init__.py
@@ -1,6 +1,7 @@
 from .emlp import EMLP, MLP  # noqa: D104
 from .imlp import IMLP
-from .time_cnn_encoder import TimeCNNEncoder, eTimeCNNEncoder
+from .time_cnn.cnn_encoder import TimeCNNEncoder
+from .time_cnn.ecnn_encoder import eTimeCNNEncoder
 
 __all__ = [
     "EMLP",
